Use remote WAR

+++ core/deploy.sh
@@ -2,48 +2,45 @@
 
 /etc/init.d/ssh start
 
-#echo Starting "$WAR_URL" 
+echo Starting "$WAR_URL" 
 #Copy the war file from mounted directory to tomcat webapps directory
-#if [ ! -z "$WAR_URL" ]
-#then
-rm -r /var/lib/tomcat7/webapps/*
-cd /var/lib/tomcat7/webapps/  #&& wget -q $WAR_URL -O ROOT.war
+if [ ! -z "$WAR_URL" ]
+then
+	rm -r /var/lib/tomcat7/webapps/*
+	cd /var/lib/tomcat7/webapps/  && wget -q $WAR_URL -O ROOT.war
 
-cp /tmp/ROOT.war /var/lib/tomcat7/webapps/ROOT.war
-chmod 777 /var/lib/tomcat7/webapps/ROOT.war
+	mkdir -p /usr/share/tomcat7/.grails
+	cd /usr/share/tomcat7/.grails
 
-mkdir -p /usr/share/tomcat7/.grails
-cd /usr/share/tomcat7/.grails
+	cp /tmp/cytomineconfig.groovy ./
 
-cp /tmp/cytomineconfig.groovy ./
+	### transform the ims urls for the config file ###
+	arr=$(echo $IMS_URLS | tr "," "\n")
+	arr=$(echo $arr | tr "[" "\n")
+	arr=$(echo $arr | tr "]" "\n")
 
-### transform the ims urls for the config file ###
-arr=$(echo $IMS_URLS | tr "," "\n")
-arr=$(echo $arr | tr "[" "\n")
-arr=$(echo $arr | tr "]" "\n")
+	IMS_URLS="["
+	for x in $arr
+	do
+	    IMS_URLS="${IMS_URLS}'http://$x',"
+	done
+	IMS_URLS="${IMS_URLS%?}"
+	IMS_URLS="$IMS_URLS]"
 
-IMS_URLS="["
-for x in $arr
-do
-    IMS_URLS="${IMS_URLS}'http://$x',"
-done
-IMS_URLS="${IMS_URLS%?}"
-IMS_URLS="$IMS_URLS]"
+	### END transform the ims urls for the config file ###
 
-### END transform the ims urls for the config file ###
+	echo "grails.serverURL='http://$CORE_URL'" >> cytomineconfig.groovy
+	echo "storage_buffer='$IMS_BUFFER_PATH'" >> cytomineconfig.groovy
+	echo "storage_path='$IMS_STORAGE_PATH'" >> cytomineconfig.groovy
 
-echo "grails.serverURL='http://$CORE_URL'" >> cytomineconfig.groovy
-echo "storage_buffer='$IMS_BUFFER_PATH'" >> cytomineconfig.groovy
-echo "storage_path='$IMS_STORAGE_PATH'" >> cytomineconfig.groovy
+	echo "grails.imageServerURL=$IMS_URLS" >> cytomineconfig.groovy
+	echo "grails.retrievalServerURL =['http://$RETRIEVAL_URL']" >> cytomineconfig.groovy
+	echo "grails.mongo.host = 'mongodb'" >> cytomineconfig.groovy
 
-echo "grails.imageServerURL=$IMS_URLS" >> cytomineconfig.groovy
-echo "grails.retrievalServerURL =['http://$RETRIEVAL_URL']" >> cytomineconfig.groovy
-echo "grails.mongo.host = 'mongodb'" >> cytomineconfig.groovy
-
-echo "grails.uploadURL='http://$UPLOAD_URL:81'" >> cytomineconfig.groovy
+	echo "grails.uploadURL='http://$UPLOAD_URL:81'" >> cytomineconfig.groovy
 
 
-#fi
+fi
 
 if [ $IS_LOCAL = true ]; then
 	echo "#Custom adding" >> /etc/hosts
