+++ ims/deploy.sh
@@ -21,29 +21,26 @@ chown -R tomcat7:tomcat7 $IMS_STORAGE_PATH
 
 export LD_LIBRARY_PATH=/usr/local/lib/openslide-java
 
-#echo Starting "$WAR_URL" 
+echo Starting "$WAR_URL" 
 #Copy the war file from mounted directory to tomcat webapps directory
-#if [ ! -z "$WAR_URL" ]
-#then
-rm -r /var/lib/tomcat7/webapps/*
-cd /var/lib/tomcat7/webapps/ # && wget -q $WAR_URL -O ROOT.war
+if [ ! -z "$WAR_URL" ]
+then
+	rm -r /var/lib/tomcat7/webapps/*
+	cd /var/lib/tomcat7/webapps/ && wget -q $WAR_URL -O ROOT.war
 
-cp /tmp/ROOT.war ./ROOT.war
-chmod 777 /var/lib/tomcat7/webapps/ROOT.war
-
-mkdir -p /usr/share/tomcat7/.grails
-cd /usr/share/tomcat7/.grails
-touch imageserverconfig.properties
-echo "dataSource.url=jdbc:h2:/tmp/devDb;MVCC=TRUE;LOCK_TIMEOUT=10000" >> imageserverconfig.properties
-echo "cytomine.storageBufferPath=$IMS_BUFFER_PATH" >> imageserverconfig.properties
-echo "cytomine.imageServerPublicKey=4a5c7004-b6f8-4705-a118-c15d5c90dcdb" >> imageserverconfig.properties
-echo "cytomine.imageServerPrivateKey=70f35a45-c317-405a-8056-353db3d2bf56" >> imageserverconfig.properties
-echo "cytomine.vips=/usr/local/bin/vips" >> imageserverconfig.properties
-echo "cytomine.identify=identify" >> imageserverconfig.properties
-echo "cytomine.tiffinfo=tiffinfo" >> imageserverconfig.properties
-echo "cytomine.vipsthumbnail=/usr/local/bin/vipsthumbnail" >> imageserverconfig.properties
-echo "cytomine.iipImageServer=http://$IIP_URL:81/fcgi-bin/iipsrv.fcgi" >> imageserverconfig.properties
-#fi
+	mkdir -p /usr/share/tomcat7/.grails
+	cd /usr/share/tomcat7/.grails
+	touch imageserverconfig.properties
+	echo "dataSource.url=jdbc:h2:/tmp/devDb;MVCC=TRUE;LOCK_TIMEOUT=10000" >> imageserverconfig.properties
+	echo "cytomine.storageBufferPath=$IMS_BUFFER_PATH" >> imageserverconfig.properties
+	echo "cytomine.imageServerPublicKey=4a5c7004-b6f8-4705-a118-c15d5c90dcdb" >> imageserverconfig.properties
+	echo "cytomine.imageServerPrivateKey=70f35a45-c317-405a-8056-353db3d2bf56" >> imageserverconfig.properties
+	echo "cytomine.vips=/usr/local/bin/vips" >> imageserverconfig.properties
+	echo "cytomine.identify=identify" >> imageserverconfig.properties
+	echo "cytomine.tiffinfo=tiffinfo" >> imageserverconfig.properties
+	echo "cytomine.vipsthumbnail=/usr/local/bin/vipsthumbnail" >> imageserverconfig.properties
+	echo "cytomine.iipImageServer=http://$IIP_URL:81/fcgi-bin/iipsrv.fcgi" >> imageserverconfig.properties
+fi
 service tomcat7 start
 
 echo "/var/log/tomcat7/catalina.out {"   > /etc/logrotate.d/tomcat7
